monolith/src/main/java/org/springframework/samples/petclinic/controller/ManagementController.java
package org.springframework.samples.petclinic.controller;

import java.util.Map;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.samples.petclinic.managementdto.ManagementClient;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.GetMapping;

@Controller
public class ManagementController {

    @Autowired
    private ManagementClient managementClient;

    @GetMapping("/management/revenue")
    public String showRevenue(Map<String, Object> model) {
        model.put("revenues", managementClient.listYearlyRevenue());
        return "management/showRevenue";
    }

}
